refactor: Separate config and CLI, unify entry point
- Moves static ticker symbol lists and other configurations from the main script to a dedicated `src/stock_analysis/config.py`.
- Relocates the `setup_arg_parser` function to a new `src/stock_analysis/cli.py` module to handle command-line interface logic.
- Renames the primary executable script from `dynamic.py` to `run.py` to better reflect its purpose.
- Deletes the unused and placeholder `main.py` file.
- Updates the main script (`run.py`) to import the configuration and CLI functions from their new, modular locations.

This refactoring improves the project's structure by enhancing modularity and separation of concerns, making the codebase easier to maintain and understand.
