+++ src/stock_analysis/config.py
@@ -1,0 +1,62 @@
+# --- 您可以在這裡修改共用參數 (You can modify parameters here) ---
+# 持有小時 (Holding Hours) - 兩項分析共用
+# 股票代碼 (Ticker Symbol) - 兩項分析共用
+TICKER_SYMBOLS_US_RARE_EARTH = ['MP','UUUU','UAMY']
+TICKER_SYMBOLS_US_DRONE = ['ONDS','RCAT'] #AVAV
+TICKER_SYMBOLS_US_NUCLEAR = ['LEU','SMR']
+TICKER_SYMBOLS_US_POWER = ['BE','VST']
+TICKER_SYMBOLS_US_BETTERY = ['EOSE','WWR']
+TICKER_SYMBOLS_US_AI_UP_1 = ['AMD','NVDA','TSM']
+TICKER_SYMBOLS_US_AI_UP_2 = ['ALAB','NVTS','POWI']
+TICKER_SYMBOLS_US_AI_MEDIUM = ['GOOG']
+TICKER_SYMBOLS_US_AI_DOWN_1 = ['ADBE','DUOL','FIG']
+TICKER_SYMBOLS_US_AI_DOWN_2 = ['GRAB','RBRK']
+
+TICKER_SYMBOLS_US_OTHER_STABLE = ['CIFR','SOFI','IBIT']
+TICKER_SYMBOLS_US_OTHER_GROWTH = ['IONQ','TMDX','TSLA']
+TICKER_SYMBOLS_US_OTHER_ETF = ['BND','GLD','MGK','SIVR','VOO']
+
+TICKER_SYMBOLS_US_OTHER = []
+
+TICKER_LIST_ARRAY = [
+    TICKER_SYMBOLS_US_RARE_EARTH,
+    TICKER_SYMBOLS_US_DRONE,
+    TICKER_SYMBOLS_US_OTHER,
+    TICKER_SYMBOLS_US_NUCLEAR,
+    TICKER_SYMBOLS_US_POWER,
+    TICKER_SYMBOLS_US_BETTERY,
+    TICKER_SYMBOLS_US_AI_UP_1,
+    TICKER_SYMBOLS_US_AI_UP_2,
+    TICKER_SYMBOLS_US_AI_MEDIUM,
+    TICKER_SYMBOLS_US_AI_DOWN_1,
+    TICKER_SYMBOLS_US_AI_DOWN_2,
+    TICKER_SYMBOLS_US_OTHER_STABLE,
+    TICKER_SYMBOLS_US_OTHER_GROWTH,
+    TICKER_SYMBOLS_US_OTHER_ETF
+]
+
+TICKER_SYMBOLS_US = TICKER_SYMBOLS_US_RARE_EARTH + \
+    TICKER_SYMBOLS_US_DRONE + \
+    TICKER_SYMBOLS_US_OTHER + \
+    TICKER_SYMBOLS_US_NUCLEAR + \
+    TICKER_SYMBOLS_US_POWER + \
+    TICKER_SYMBOLS_US_BETTERY + \
+    TICKER_SYMBOLS_US_AI_UP_1 + \
+    TICKER_SYMBOLS_US_AI_UP_2 + \
+    TICKER_SYMBOLS_US_AI_MEDIUM + \
+    TICKER_SYMBOLS_US_AI_DOWN_1 + \
+    TICKER_SYMBOLS_US_AI_DOWN_2 + \
+    TICKER_SYMBOLS_US_OTHER_STABLE + \
+    TICKER_SYMBOLS_US_OTHER_GROWTH + \
+    TICKER_SYMBOLS_US_OTHER_ETF
+
+TICKER_SYMBOLS_TW = ['00635U.TW','2603.TW']
+TICKER_SYMBOLS = TICKER_SYMBOLS_US
+
+
+# --- 參數定義 (Parameter Definitions) ---
+# (保留儲存格 5, 6, 7/8 中的所有參數定義)
+
+# 分析 2 & 3 參數
+INTERVAL_LONG = '60m'
+# (註：原儲存格 7 和 7/8 都使用 '60m' 和 '5d'，我們在批次下載時使用)
